webserver/requirements.txt
flask-cors
pillow>=11.0
pymilvus>=2.6.0
marshmallow==3.21.0
flask
flask_restful
gunicorn>=23.0.0
grpcio<2.0.0
protobuf<6,>=3.20
blinker==1.4
# PyTorch already provided by base image
# Transformers for CLIP model
transformers>=4.57.0
accelerate
# google model need
sentencepiece
einops
timm